fix: enable fact list scroll
