+++ src/app/layout.tsx
@@ -29,7 +29,7 @@ export default function RootLayout({
     return (
         <html lang="en">
             <body
-                className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+                className={`${geistSans.variable} ${geistMono.variable} pt-14 antialiased`}
             >
                 <Header />
                 <MantineProvider
