+++ src/components/AppShell/Header.tsx
@@ -1,10 +1,10 @@
 export default function Header() {
     return (
         <>
-            <div className="fixed left-0 right-0 top-0 flex h-14 items-center justify-center bg-neutral-100">
+            <div className="fixed left-0 right-0 top-0 z-20 flex h-14 items-center justify-center bg-neutral-100">
                 <h1 className="text-2xl font-bold text-black">CommonGround</h1>
             </div>
-            <div className="h-14 bg-transparent"></div>
+            {/* <div className="h-14 bg-transparent"></div> */}
         </>
     );
 }
